+++ index.ts
@@ -65,3 +65,25 @@ export type {
 	TransactWriteOptions,
 	TransactWriteResult,
 } from './lib/transact.js';
+
+export type {
+	BeginsWithCondition,
+	BetweenCondition,
+	Comparator,
+	ComparatorCondition,
+	Condition,
+	ConditionExpression,
+	ContainsCondition,
+	ExistsCondition,
+	FilterCondition,
+	FilterConditionExpression,
+	FilterLogicEvaluation,
+	InCondition,
+	LogicEvaluation,
+	NotExistsCondition,
+	NotExpression,
+	SizeCondition,
+} from './lib/expression/condition.js';
+
+export type { AttributeMap } from './lib/converter/converter.js';
+export type { ConverterOptions } from './lib/converter/converter-options.js';